Restructured for Compilation
